+++ packages/transpiler/package.json
@@ -16,6 +16,7 @@
   },
   "devDependencies": {
     "@near-js/types": "^0.0.4",
+    "@types/babel__standalone": "^7.1.4",
     "@types/node": "^17.0.12",
     "eslint": "^7.32.0",
     "eslint-config-custom": "workspace:*",
