+++ packages/transpiler/src/transpiler.ts
@@ -1,0 +1,13 @@
+import Babel from '@babel/standalone';
+
+/**
+ * Transpile the Component's JSX source code, replacing React.createElement with `createElement`
+ * @param source Component function source code using JSX
+ */
+export function transpileSource(source: string) {
+  return Babel.transform(source, {
+    plugins: [
+      [Babel.availablePlugins['transform-react-jsx'], { pragma: 'createElement' }],
+    ],
+  });
+}
